+++ src/misc/types.ts
@@ -7,7 +7,6 @@ import {Positioner} from "../plugins/positioner";
 import {ReadonlyState} from "../plugins/readonlyState";
 import {ValueHandler} from "../plugins/valueHandler";
 import {LiveSearch} from "../plugins/liveSearch";
-import {TextsLocator} from "../plugins/textsLocator";
 import {NgSelectOptions} from "./ngSelectOptions.interface";
 
 /**
@@ -49,8 +48,3 @@ export const VALUE_HANDLER_TYPE: InjectionToken<Type<ValueHandler<any>>> = new I
  * Injection token for 'LiveSearch' implementation
  */
 export const LIVE_SEARCH_TYPE: InjectionToken<Type<LiveSearch>> = new InjectionToken<Type<LiveSearch>>('LIVE_SEARCH_TYPE');
-
-/**
- * Injection token for 'TextsLocator' implementation
- */
-export const TEXTS_LOCATOR_TYPE: InjectionToken<Type<TextsLocator>> = new InjectionToken<Type<TextsLocator>>('TEXTS_LOCATOR_TYPE');