+++ src/plugins/textsLocator/index.ts
@@ -1,3 +1,0 @@
-export * from './textsLocator.interface';
-
-export * from './no/noTextsLocator.interface';